@@ -42,7 +42,8 @@ const COUNTRIES: Country[] = [
 export type SortColumn = keyof Country | '';
 export type SortDirection = 'asc' | 'desc' | '';
 const rotate: {[key: string]: SortDirection} = { 'asc': 'desc', 'desc': '', '': 'asc' };
-const compare = (v1: string, v2: string) => v1 < v2 ? -1 : v1 > v2 ? 1 : 0;
+
+const compare = (v1: string | number, v2: string | number) => v1 < v2 ? -1 : v1 > v2 ? 1 : 0;
 
 export interface SortEvent {
   column: SortColumn;
@@ -93,7 +94,7 @@ export class NgbdTableSortable {
       this.countries = COUNTRIES;
     } else {
       this.countries = [...COUNTRIES].sort((a, b) => {
-        const res = compare(`${a[column]}`, `${b[column]}`);
+        const res = compare(a[column], b[column]);
         return direction === 'asc' ? res : -res;
       });
     }
