@@ -5,7 +5,7 @@ import {Component, Input} from '@angular/core';
   templateUrl: './icons.component.html',
   host: {
     'xmlns': 'http://www.w3.org/2000/svg',
-    'viewBox': '0 0 24 24',
+    '[attr.viewBox]': '"0 0 "+width+" "+height',
     '[attr.width]': 'width',
     '[attr.height]': 'height',
   }
