@@ -1,8 +1,8 @@
+let bootstrap: string = require('../../../package.json').devDependencies['bootstrap'];
 // extracts only the minor version from package.json
 // ex. "bootstrap": "4.0.1" -> "4.0"
-let bootstrap: string = require('../../../package.json').devDependencies['bootstrap'];
-bootstrap = bootstrap.substr(0, bootstrap.lastIndexOf('.'));
+bootstrap = bootstrap.split('.').slice(0, 2).join('.');
 
-const ngBootstrap = require('../../../src/package.json').version;
+const ngBootstrap = require('../../../package.json').version;
 
-export const versions: {[key: string]: string} = { bootstrap, ngBootstrap };
+export const versions: {[key: string]: string} = {bootstrap, ngBootstrap};
