@@ -1,12 +1,11 @@
 import * as path from 'path';
-import {Rule, SchematicContext, Tree} from '@angular-devkit/schematics';
-import {getWorkspace} from '@schematics/angular/utility/config';
-import {getProject} from '@schematics/angular/utility/project';
-import {WorkspaceProject, WorkspaceSchema} from '@schematics/angular/utility/workspace-models';
+import {Rule, SchematicContext, Tree, SchematicsException} from '@angular-devkit/schematics';
 
 import {Schema} from '../schema';
 import * as messages from '../messages';
 import {getProjectStyleFile, getProjectTargetOptions} from '../../utils/project';
+import {getWorkspace, updateWorkspace} from '@schematics/angular/utility/workspace';
+import {workspaces, JsonArray} from '@angular-devkit/core';
 
 
 const BOOTSTRAP_CSS_FILEPATH = 'node_modules/bootstrap/dist/css/bootstrap.min.css';
@@ -26,16 +25,22 @@ const SUPPORTED_BOOTSTRAP_STYLE_IMPORTS = {
  * If not possible, we're simply adding 'bootstrap.css' to the 'angular.json'
  */
 export function addBootstrapStyles(options: Schema): Rule {
-  return (host: Tree, context: SchematicContext) => {
-    const workspace = getWorkspace(host);
-    const project = getProject(workspace, options.project || workspace.defaultProject !);
+  return async(host: Tree, context: SchematicContext) => {
+    const workspace = await getWorkspace(host);
+
+    const projectName = options.project || workspace.extensions.defaultProject !.toString();
+    const project = workspace.projects.get(projectName);
+    if (!project) {
+      throw new SchematicsException(messages.noProject(projectName));
+    }
+
     const styleFilePath = getProjectStyleFile(project) || '';
     const styleFileExtension = path.extname(styleFilePath);
     const styleFilePatch = SUPPORTED_BOOTSTRAP_STYLE_IMPORTS[styleFileExtension];
 
     // found supported styles
     if (styleFilePatch) {
-      addBootstrapToStylesFile(host, styleFilePath, styleFilePatch);
+      return addBootstrapToStylesFile(styleFilePath, styleFilePatch);
     } else {
       // found some styles, but unsupported
       if (styleFileExtension !== '.css' && styleFileExtension !== '') {
@@ -43,40 +48,45 @@ export function addBootstrapStyles(options: Schema): Rule {
       }
 
       // just patching 'angular.json'
-      addBootstrapToAngularJson(workspace, project, host);
+      return addBootstrapToAngularJson(workspace, project, host);
     }
-    return host;
   };
 }
 
 /**
  * Patches 'styles.scss' or 'styles.sass' to add Bootstrap snippet
  */
-function addBootstrapToStylesFile(host: Tree, styleFilePath: string, styleFilePatch: string) {
-  const styleContent = host.read(styleFilePath) !.toString('utf-8');
+function addBootstrapToStylesFile(styleFilePath: string, styleFilePatch: string): Rule {
+  return (host: Tree) => {
+    const styleContent = host.read(styleFilePath) !.toString('utf-8');
 
-  const recorder = host.beginUpdate(styleFilePath);
-  recorder.insertRight(styleContent.length, styleFilePatch);
+    const recorder = host.beginUpdate(styleFilePath);
+    recorder.insertRight(styleContent.length, styleFilePatch);
 
-  host.commitUpdate(recorder);
+    host.commitUpdate(recorder);
+  };
 }
 
 /**
  * Patches 'angular.json' to add 'bootstrap.css' styles
  */
-function addBootstrapToAngularJson(workspace: WorkspaceSchema, project: WorkspaceProject, host: Tree) {
+function addBootstrapToAngularJson(
+    workspace: workspaces.WorkspaceDefinition, project: workspaces.ProjectDefinition, host: Tree): Rule {
   const targetOptions = getProjectTargetOptions(project, 'build');
-  if (!targetOptions.styles) {
+  const styles = (targetOptions.styles as JsonArray | undefined);
+  if (!styles) {
     targetOptions.styles = [BOOTSTRAP_CSS_FILEPATH];
   } else {
-    const existingStyles = targetOptions.styles.map((s) => typeof s === 'string' ? s : s.input);
+    const existingStyles = styles.map((s) => typeof s === 'string' ? s : s !['input']);
+
     for (const[, stylePath] of existingStyles.entries()) {
       // If the given asset is already specified in the styles, we don't need to do anything.
       if (stylePath === BOOTSTRAP_CSS_FILEPATH) {
-        return;
+        return () => host;
       }
     }
-    targetOptions.styles.unshift(BOOTSTRAP_CSS_FILEPATH);
+    styles.unshift(BOOTSTRAP_CSS_FILEPATH);
   }
-  host.overwrite('angular.json', JSON.stringify(workspace, null, 2));
+
+  return updateWorkspace(workspace);
 }
