@@ -1,6 +1,4 @@
 import {Rule, SchematicsException, Tree} from '@angular-devkit/schematics';
-import {getWorkspace} from '@schematics/angular/utility/config';
-import {getProject} from '@schematics/angular/utility/project';
 import {getAppModulePath} from '@schematics/angular/utility/ng-ast-utils';
 import {addImportToModule} from '@schematics/angular/utility/ast-utils';
 import {InsertChange} from '@schematics/angular/utility/change';
@@ -8,6 +6,8 @@ import * as ts from '@schematics/angular/third_party/github.com/Microsoft/TypeSc
 
 import {Schema} from '../schema';
 import {getProjectTargetOptions} from '../../utils/project';
+import {getWorkspace} from '@schematics/angular/utility/workspace';
+import * as messages from '../messages';
 
 
 const NG_BOOTSTRAP_MODULE_NAME = 'NgbModule';
@@ -17,12 +17,16 @@ const NG_BOOTSTRAP_PACKAGE_NAME = '@ng-bootstrap/ng-bootstrap';
  * Patches main application module by adding 'NgbModule' import
  */
 export function addNgbModuleToAppModule(options: Schema): Rule {
-  return (host: Tree) => {
-    const workspace = getWorkspace(host);
-    const project = getProject(workspace, options.project || workspace.defaultProject !);
+  return async(host: Tree) => {
+    const workspace = await getWorkspace(host);
+    const projectName = options.project || (workspace.extensions.defaultProject as string);
+    const project = workspace.projects.get(projectName);
+    if (!project) {
+      throw new SchematicsException(messages.noProject(projectName));
+    }
     const buildOptions = getProjectTargetOptions(project, 'build');
 
-    const modulePath = getAppModulePath(host, buildOptions.main);
+    const modulePath = getAppModulePath(host, (buildOptions.main as string));
 
     const text = host.read(modulePath);
     if (text === null) {
@@ -40,6 +44,5 @@ export function addNgbModuleToAppModule(options: Schema): Rule {
       }
     }
     host.commitUpdate(recorder);
-    return host;
   };
 }
