@@ -1,6 +1,4 @@
-import {normalize} from '@angular-devkit/core';
-import {WorkspaceProject} from '@schematics/angular/utility/workspace-models';
-import {getProjectTargets} from '@schematics/angular/utility/project-targets';
+import {normalize, workspaces} from '@angular-devkit/core';
 import {SchematicsException} from '@angular-devkit/schematics';
 
 // Regular expression that matches all possible Angular CLI default style files
@@ -12,10 +10,10 @@ const validStyleFileRegex = /\.(c|le|sc|sa)ss/;
 /**
  * Resolves options for the build target of the given project
  */
-export function getProjectTargetOptions(project: WorkspaceProject, buildTarget: string) {
-  const targets = getProjectTargets(project);
-  if (targets && targets[buildTarget] && targets[buildTarget].options) {
-    return targets[buildTarget].options;
+export function getProjectTargetOptions(project: workspaces.ProjectDefinition, buildTarget: string) {
+  const buildTargetObject = project.targets.get(buildTarget);
+  if (buildTargetObject && buildTargetObject.options) {
+    return buildTargetObject.options;
   }
 
   throw new SchematicsException(`Cannot determine project target configuration for: ${buildTarget}.`);
@@ -26,11 +24,11 @@ export function getProjectTargetOptions(project: WorkspaceProject, buildTarget: 
  * Gets a style file with the given extension in a project and returns its path. If no
  * extension is specified, any style file with a valid extension will be returned.
  */
-export function getProjectStyleFile(project: WorkspaceProject, extension?: string): string | null {
+export function getProjectStyleFile(project: workspaces.ProjectDefinition, extension?: string): string | null {
   const buildOptions = getProjectTargetOptions(project, 'build');
 
-  if (buildOptions.styles && buildOptions.styles.length) {
-    const styles = buildOptions.styles.map((s) => typeof s === 'string' ? s : s.input);
+  if (buildOptions.styles && Array.isArray(buildOptions.styles) && buildOptions.styles.length) {
+    const styles = buildOptions.styles.map((s) => typeof s === 'string' ? s : s !['input']);
 
     // Look for the default style file that is generated for new projects by the Angular CLI. This
     // default style file is usually called `styles.ext` unless it has been changed explicitly.
