@@ -66,8 +66,21 @@ if (typeof Element !== 'undefined' && !Element.prototype.closest) {
   };
 }
 
-export function closest(element: HTMLElement, selector): HTMLElement | null {
+export function closest(element: HTMLElement, selector?: string): HTMLElement | null {
   if (!selector) {
+    return null;
+  }
+
+  /*
+   * In certain browsers (e.g. Edge 44.18362.449.0) HTMLDocument does
+   * not support `Element.prototype.closest`. To emulate the correct behaviour
+   * we return null when the method is missing.
+   *
+   * Note that in evergreen browsers `closest(document.documentElement, 'html')`
+   * will return the document element whilst in Edge null will be returned. This
+   * compromise was deemed good enough.
+   */
+  if (typeof element.closest === 'undefined') {
     return null;
   }
 
